requirements.txt
Package                 Version
----------------------- -------
colorama                0.4.6
customtkinter           5.2.2
darkdetect              0.8.0
Deprecated              1.2.14
img2pdf                 0.5.1
lxml                    5.3.0
Markdown                3.7
packaging               24.1
pikepdf                 9.4.0
pillow                  11.0.0
pip                     24.2
py-mon                  2.0.5
PyPDF4                  1.27.0
pyqt-toast-notification 1.3.2
PyQt5                   5.15.11
PyQt5-Qt5               5.15.15
PyQt5_sip               12.15.0
PyQt6                   6.7.1
PyQt6-Qt6               6.7.3
PyQt6_sip               13.8.0
pyqtspinner             2.0.0
pytesseract             0.3.13
python-dotenv           1.0.1
QtPy                    2.4.2
qtwidgets               1.1
watchdog                6.0.0
wrapt                   1.16.0
